Move grok_pattern.go away from pkg/types to reduce bouncer dependencies
The right place seems pkg/parser, and all files using grok_pattern
already import it.
